requirements.txt
requests
pandas
pyarrow
PyYAML
python-dotenv
joblib
pytest
ruff
black
